review fixes

+++ contracts/EVMScriptFactories/ActivateNodeOperators.sol
@@ -74,7 +74,7 @@ contract ActivateNodeOperators is TrustedCaller, IEVMScriptFactory {
         bytes4[] memory methodIds = new bytes4[](decodedCallData.length * 2);
         bytes[] memory encodedCalldata = new bytes[](decodedCallData.length * 2);
 
-        for (uint i = 0; i < decodedCallData.length; i++) {
+        for (uint256 i = 0; i < decodedCallData.length; i++) {
             toAddresses[i * 2] = address(nodeOperatorsRegistry);
             methodIds[i * 2] = ACTIVATE_NODE_OPERATOR_SELECTOR;
             encodedCalldata[i * 2] = abi.encode(decodedCallData[i].nodeOperatorId);
@@ -115,7 +115,7 @@ contract ActivateNodeOperators is TrustedCaller, IEVMScriptFactory {
         ActivateNodeOperatorInput[] memory _activateNodeOperatorInputs
     ) private view {
         uint256 nodeOperatorsCount = nodeOperatorsRegistry.getNodeOperatorsCount();
-        for (uint i = 0; i < _activateNodeOperatorInputs.length; i++) {
+        for (uint256 i = 0; i < _activateNodeOperatorInputs.length; i++) {
             require(
                 _activateNodeOperatorInputs[i].nodeOperatorId < nodeOperatorsCount,
                 NODE_OPERATOR_INDEX_OUT_OF_RANGE
