+++ contracts/EVMScriptFactories/SetNodeOperatorsName.sol
@@ -55,7 +55,7 @@ contract SetNodeOperatorNames is TrustedCaller, IEVMScriptFactory {
 
         bytes[] memory nodeOperatorsNamesCalldata = new bytes[](decodedCallData.length);
 
-        for (uint i = 0; i < decodedCallData.length; i++) {
+        for (uint256 i = 0; i < decodedCallData.length; i++) {
             nodeOperatorsNamesCalldata[i] = abi.encode(
                 decodedCallData[i].nodeOperatorId,
                 decodedCallData[i].name
@@ -88,7 +88,7 @@ contract SetNodeOperatorNames is TrustedCaller, IEVMScriptFactory {
 
     function _validateInputData(SetNameInput[] memory _nodeOperatorNamesInput) private view {
         uint256 nodeOperatorsCount = nodeOperatorsRegistry.getNodeOperatorsCount();
-        for (uint i = 0; i < _nodeOperatorNamesInput.length; i++) {
+        for (uint256 i = 0; i < _nodeOperatorNamesInput.length; i++) {
             require(
                 _nodeOperatorNamesInput[i].nodeOperatorId < nodeOperatorsCount,
                 NODE_OPERATOR_INDEX_OUT_OF_RANGE
